pyproject.toml
[tool.poetry]
name = "micro-paas"
version = "0.1.0"
description = ""
authors = ["wangyouzheng <353735348@qq.com>"]
readme = "readme.md"
# 把project
packages = [{ include = "project", from = "src" }]


[tool.poetry.dependencies]
python = "^3.11"
click = "^8.1.7"
PyYAML = "^6.0.2"
requests = "^2.32.3"
pymysql = "^1.1.1"
redis = "^5.0.8"
cx_Oracle = { version = ">=8.3.0", optional = true }
psutil = "^7.0.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
monitor = "project.cli:cli"